requirements.txt
absl-py==2.0.0
addict==2.4.0
albumentations==1.3.1
asgiref==3.7.2
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
coloredlogs==15.0.1
contourpy==1.1.1
customtkinter==5.2.0
cycler==0.12.0
Cython==3.0.2
darkdetect==0.8.0
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
easydict==1.10
facexlib==0.3.0
ffmpeg==1.4
filelock==3.12.4
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.43.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
gfpgan==1.3.8
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
humanfriendly==10.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
insightface==0.7.3
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.0
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.24.3
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
opennsfw2==0.10.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.0.0
platformdirs==3.11.0
prettytable==3.9.0
protobuf==4.23.4
psutil==5.9.5
psycopg==3.1.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tk==0.1.0
tkinterdnd2==0.3.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.5
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.15.0
yapf==0.40.2
zipp==3.17.0
